AMP-2927
git-svn-id: file:///svn/trunk@4721 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/action/AddSelectedSectorsForInd.java
@@ -42,11 +42,11 @@ public class AddSelectedSectorsForInd extends Action {
 							logger.info("adding now...");
                             if(eaForm.getActivitySectors()==null){
                                 if(selsearchedSector.length==1){
-                                    sctr.setSectorPercentage(new Integer(100));
+                                    sctr.setSectorPercentage(new Float(100));
                                 }
                             }else if(eaForm.getActivitySectors().size()==0){
                                 if(selsearchedSector.length==1){
-                                    sctr.setSectorPercentage(new Integer(100));
+                                    sctr.setSectorPercentage(new Float(100));
                                 }
                             }else if(eaForm.getActivitySectors().size()==1){
                                 Iterator sectItr=eaForm.getActivitySectors().iterator();
