+++ amp/WEB-INF/src/org/digijava/module/aim/action/SectorSelectedForIndicator.java
@@ -91,9 +91,9 @@ public class SectorSelectedForIndicator extends Action {
 			}
 
             if(prevSelSectors==null){
-                actSect.setSectorPercentage(new Integer(100));
+                actSect.setSectorPercentage(new Float(100));
             }else if(prevSelSectors.size()==0){
-                    actSect.setSectorPercentage(new Integer(100));
+                    actSect.setSectorPercentage(new Float(100));
             }else if(prevSelSectors.size()==1){
                 Iterator prevItr = prevSelSectors.iterator();
                 while(prevItr.hasNext()) {
