+++ amp/WEB-INF/src/org/digijava/module/aim/action/ShowActivityPrintPreview.java
@@ -965,7 +965,7 @@ public class ShowActivityPrintPreview
             									actCompo.setSubsectorLevel2Name(subsectorLevel2.getName());
             								}
             							}
-            							actCompo.setSectorPercentage(ampActSect.getPercentage().intValue());
+            							actCompo.setSectorPercentage(ampActSect.getPercentage().floatValue());
             						}
             						activitySectors.add(actCompo);
             					}
